feat: add concurrent test running to littletest
BREAKING CHANGE: TestEvent objects now contain TestInfo, DescribeInfo, and
LifecycleInfo object instead of Test, Describe, and Lifecycle objects. Also,
instead of including Error objects, they include JSON-compatible error-like
objects (with a name, message, and stack property, as well as any additional
properties that were on the error). This makes TestEvents JSON-serializable.

BREAKING CHANGE: the "flags" property on Test, Describe, and Lifecycle objects
is now an Array instead of a Set. This is to make it consistent with the new
TestInfo, DescribeInfo, and LifecycleInfo objects, which have to be
JSON-serializable.

BREAKING CHANGE: the "run_finished" event's `events` property no longer includes
the "run_finished" event itself.